backend/gcp_function/requirements.txt
functions-framework==3.0.0
google-cloud-storage==2.1.0
